pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-daemon"
version = "0.1.0"
description = "Local macOS push-to-talk daemon — hotkey → record → transcribe → paste"
requires-python = ">=3.11"
dependencies = [
    "mlx-whisper>=0.4",
    "sounddevice>=0.5",
    "numpy>=1.24",
    "pynput>=1.7",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "onnxruntime>=1.16",
    "rumps>=0.4",
]

[project.scripts]
whisper-daemon = "whisper_daemon.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
